+++ pipfreeze/pipfreeze.py
@@ -7,10 +7,17 @@ class PipFreeze(object):
 
     def _load_pip_freeze(self, pip_freeze_output):
         self._requirements = {}
+        self._order = []
 
         reqs = pkg_resources.parse_requirements(pip_freeze_output)
         for req in reqs:
             self._requirements[req.key] = req
+            self._order.append(req.key)
+
+    def __iter__(self):
+        for key in self._order:
+            req = self._requirements[key]
+            yield req
 
     def __nonzero__(self):
         # Python 2.x