+++ CHANGES.rst
@@ -2,6 +2,7 @@ Unreleased
 ==========
 
 * implement ``PipFreeze.__bool__()`` (py3) and ``PipFreeze.__nonzero__()`` (py2)
+* implement ``PipFreeze.__iter__()``
 
 0.0.1
 =====